refactor: renamed common.dart, preface proto class'
