+++ lib/app_center_ratings_client.dart
@@ -3,18 +3,18 @@ import 'dart:async';
 import 'package:grpc/grpc.dart';
 import 'package:meta/meta.dart';
 
-import 'src/chart.dart' as chart;
-import 'src/common.dart' as app;
+import 'src/chart.dart';
 import 'src/generated/google/protobuf/empty.pb.dart';
-import 'src/generated/ratings_features_app.pbgrpc.dart';
-import 'src/generated/ratings_features_chart.pbgrpc.dart';
-import 'src/generated/ratings_features_user.pbgrpc.dart';
-import 'src/user.dart' as user;
+import 'src/generated/ratings_features_app.pbgrpc.dart' as appPb;
+import 'src/generated/ratings_features_chart.pbgrpc.dart' as chartPb;
+import 'src/generated/ratings_features_user.pbgrpc.dart' as userPb;
+import 'src/ratings.dart';
+import 'src/user.dart';
 
 class RatingsClient {
-  late AppClient _appClient;
-  late UserClient _userClient;
-  late ChartClient _chartClient;
+  late appPb.AppClient _appClient;
+  late userPb.UserClient _userClient;
+  late chartPb.ChartClient _chartClient;
 
   RatingsClient(String serverUrl, int port) {
     final channel = ClientChannel(
@@ -24,9 +24,9 @@ class RatingsClient {
         credentials: ChannelCredentials.insecure(),
       ),
     );
-    _appClient = AppClient(channel);
-    _userClient = UserClient(channel);
-    _chartClient = ChartClient(channel);
+    _appClient = appPb.AppClient(channel);
+    _userClient = userPb.UserClient(channel);
+    _chartClient = chartPb.ChartClient(channel);
   }
 
   // Additional constructor for testing
@@ -38,7 +38,7 @@ class RatingsClient {
   );
 
   Future<String> authenticate(String id) async {
-    final request = AuthenticateRequest(id: id);
+    final request = userPb.AuthenticateRequest(id: id);
     final grpcResponse = await _userClient.authenticate(request);
     return grpcResponse.token;
   }
@@ -50,9 +50,8 @@ class RatingsClient {
     await _userClient.delete(request, options: callOptions);
   }
 
-  Future<List<chart.ChartData>> getChart(
-      chart.Timeframe timeframe, String token) async {
-    final request = GetChartRequest(timeframe: timeframe.toDTO());
+  Future<List<ChartData>> getChart(Timeframe timeframe, String token) async {
+    final request = chartPb.GetChartRequest(timeframe: timeframe.toDTO());
     final callOptions =
         CallOptions(metadata: {'authorization': 'Bearer $token'});
     final grpcResponse =
@@ -60,11 +59,11 @@ class RatingsClient {
     return grpcResponse.orderedChartData.map((data) => data.fromDTO()).toList();
   }
 
-  Future<app.Rating> getRating(
+  Future<Rating> getRating(
     String snapId,
     String token,
   ) async {
-    final request = GetRatingRequest(snapId: snapId);
+    final request = appPb.GetRatingRequest(snapId: snapId);
     final callOptions =
         CallOptions(metadata: {'authorization': 'Bearer $token'});
     final grpcResponse = await _appClient.getRating(
@@ -74,8 +73,8 @@ class RatingsClient {
     return grpcResponse.rating.fromDTO();
   }
 
-  Future<List<user.Vote>> getSnapVotes(String snap_id, String token) async {
-    final request = GetSnapVotesRequest(snapId: snap_id);
+  Future<List<Vote>> getSnapVotes(String snap_id, String token) async {
+    final request = userPb.GetSnapVotesRequest(snapId: snap_id);
     final callOptions =
         CallOptions(metadata: {'authorization': 'Bearer $token'});
     final grpcResponse = await _userClient.getSnapVotes(
@@ -85,8 +84,8 @@ class RatingsClient {
     return grpcResponse.votes.map((vote) => vote.fromDTO()).toList();
   }
 
-  Future<List<user.Vote>> listMyVotes(String snapIdFilter, String token) async {
-    final request = ListMyVotesRequest(snapIdFilter: snapIdFilter);
+  Future<List<Vote>> listMyVotes(String snapIdFilter, String token) async {
+    final request = userPb.ListMyVotesRequest(snapIdFilter: snapIdFilter);
     final callOptions =
         CallOptions(metadata: {'authorization': 'Bearer $token'});
     final grpcResponse = await _userClient.listMyVotes(
@@ -98,7 +97,7 @@ class RatingsClient {
 
   Future<void> vote(
       String snapId, int snapRevision, bool voteUp, String token) async {
-    final request = VoteRequest(
+    final request = userPb.VoteRequest(
       snapId: snapId,
       snapRevision: snapRevision,
       voteUp: voteUp,
